Create separate component to view trip
