+++ cypress/integration/newtrip.spec.js
@@ -200,4 +200,44 @@ describe('adding destinations', () => {
     })
   })
 
+  it('should let you create multiple trips', () => {
+    cy.fixture('pois.json').then(data => {
+      cy.intercept('GET', 'https://api.foursquare.com/v3/*', data)
+      cy.contains('Submit').click()
+      cy.get('.POICard > button').each((button) => {
+        button.click()
+      })
+      cy.contains('Home').click()
+      cy.contains('New Trip').click()
+      cy.get('.POICard:first').next().contains('Select').click()
+      cy.get('input[type=date]').click().type('1995-02-26')
+
+      cy.contains('Submit').click()
+
+      cy.get('select[name=locomotion]').should('have.value', 'walk')
+      .select('Bike')
+
+      cy.get('input[type=number]').click().clear().type('30')
+
+      cy.get('select[name=poi]')
+      .select('Grocery Stores')
+
+      cy.contains('Submit').click()
+      cy.get('.POICard').should('exist')
+      cy.get('.POICard button').each((button) => {
+        button.click()
+      })
+
+      cy.contains('Home').click()
+
+      cy.get('.TripCard').should('have.length', 2)
+
+      cy.contains('Dockside Marina').should('exist')
+      .click()
+
+      cy.get('.POICard').should('have.length', 5)
+
+    })
+  })
+
 })
